requirements.txt
gradio==5.29.0
requests
python-dotenv
